+++ core/services/persy/src/config.rs
@@ -33,10 +33,10 @@ pub struct PersyConfig {
     pub index: Option<String>,
 }
 
-impl crate::Configurator for PersyConfig {
+impl opendal_core::Configurator for PersyConfig {
     type Builder = PersyBuilder;
 
-    fn from_uri(uri: &crate::types::OperatorUri) -> crate::Result<Self> {
+    fn from_uri(uri: &opendal_core::OperatorUri) -> opendal_core::Result<Self> {
         let mut map = uri.options().clone();
 
         if let Some(path) = uri.root() {
@@ -57,8 +57,8 @@ impl crate::Configurator for PersyConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Configurator;
-    use crate::types::OperatorUri;
+    use opendal_core::Configurator;
+    use opendal_core::OperatorUri;
 
     #[test]
     fn from_uri_sets_datafile_segment_and_index() {
